user_model/requirements.txt
-i https://pypi.org/simple
asgiref==3.5.1
backports.zoneinfo==0.2.1
django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
